+++ opentelemetry-sdk/src/testing/logs/mod.rs
@@ -1,0 +1,3 @@
+pub use in_memory_exporter::{InMemoryLogsExporter, InMemoryLogsExporterBuilder};
+
+mod in_memory_exporter;